breakdown analytics page +  thread distri fix
